packages/aos_telemetry/pyproject.toml
[project]
name = "aos_telemetry"
version = "0.1.0"
description = "AOS Telemetry - OpenTelemetry setup and logger wrappers"
requires-python = ">=3.10"
dependencies = [
    "structlog>=24.0",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
]

[tool.hatch.build.targets.wheel]
packages = ["src/aos_telemetry"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
